requirements.txt
django==3.0.2
requests==2.22.0
djangorestframework==3.11.0